Refactor: Refactored Close -> CloseSvg
